backend/requirements.txt
fastapi[standard]==0.115.11
SQLAlchemy==2.0.38
asyncpg==0.30.0
alembic==1.15.1
pydantic-settings==2.8.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4